# docs/openapi/console-v1.yaml
openapi: 3.0.3
info:
  title: AI Gateway Console API
  version: "1"
  description: |
    JWT-authenticated Console API for the AI Gateway. Served on a dedicated
    listener under `/console/v1/*`, separate from the OpenAI-compatible
    `/v1/*` data plane. `admin` is a `users.role` value, not a separate
    namespace or static bearer token.

    Conventions:
    - Decimals (`rust_decimal::Decimal`) and timestamps (`chrono::DateTime<Utc>`)
      serialize as RFC 3339 / decimal strings; all such fields are `string`.
    - All monetary balances, quotas, prices, request costs, and statistics are
      denominated and settled in USD. Currency is not a configurable field.
    - Most mutable resources return an `ETag` on detail `GET` and require
      `If-Match` on `PUT` for optimistic concurrency. A stale `If-Match`
      yields `409`.
    - API key values are retrievable by their owner and administrators; the
      Console masks them by default. Invitation tokens remain one-time values
      that are never retrievable again.
    - Every error body is `{"error":"<code>"}`.
    - `routing_dependency_invalid` means a mutation would make the routing
      graph invalid (for example, by leaving an enabled model rule without an
      eligible channel). Add or retain compatible enabled resources, or update
      the affected rule first.
    - `limit` query parameters are clamped to `1..=100` (default `50`).

    Compatibility aliases (same handlers, same JWT/role boundary, not
...
servers:
...
paths:
...
  /models:
    get:
      tags: [Models]
      operationId: listModels
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Models.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ControlPlaneModel" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Models]
      operationId: createModel
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModelInput" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MutationResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Unprocessable" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Unauthorized.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
    Forbidden:
      description: Forbidden.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
...
    Unprocessable:
      description: Request rejected (validation).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
...
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    DateTime:
      type: string
      format: date-time
    Decimal:
      type: string
      description: rust_decimal::Decimal serialized as a string.
    DateTimeNullable:
      type: string
      format: date-time
      nullable: true
...
    JsonValue:
      description: Arbitrary JSON; typed as `unknown`.
      # Empty schema => openapi-typescript emits `unknown`.
...
    MutationResponse:
      type: object
      required: [id, correlation_id]
      properties:
        id: { type: string, format: uuid }
        secret:
          type: string
          description: Created API key value; present only on API key create responses.
        correlation_id: { type: string, format: uuid }
...
    ControlPlaneModel:
      type: object
      required:
        [id, source_model_id, display_name, provider_name, enabled,
         price_unit_tokens, input_unit_price, cached_input_unit_price,
         cache_write_unit_price, output_unit_price, price_effective_at,
         last_synced_at, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        source_model_id: { type: string }
        display_name: { type: string }
        provider_name: { type: string, nullable: true }
        enabled: { type: boolean }
        price_unit_tokens: { type: integer }
        input_unit_price: { $ref: "#/components/schemas/Decimal" }
        cached_input_unit_price: { $ref: "#/components/schemas/Decimal" }
        cache_write_unit_price: { $ref: "#/components/schemas/Decimal" }
        output_unit_price: { $ref: "#/components/schemas/Decimal" }
        price_effective_at: { $ref: "#/components/schemas/DateTime" }
        last_synced_at: { $ref: "#/components/schemas/DateTimeNullable" }
        created_at: { $ref: "#/components/schemas/DateTime" }
        updated_at: { $ref: "#/components/schemas/DateTime" }
...
    ModelInput:
      type: object
      required:
        [source_model_id, display_name, enabled, price_unit_tokens,
         input_unit_price, cached_input_unit_price, cache_write_unit_price,
         output_unit_price, price_effective_at]
      properties:
        source_model_id: { type: string }
        display_name: { type: string }
        provider_name: { type: string, nullable: true }
        enabled: { type: boolean }
        price_unit_tokens: { type: integer }
        input_unit_price: { $ref: "#/components/schemas/Decimal" }
        cached_input_unit_price: { $ref: "#/components/schemas/Decimal" }
        cache_write_unit_price: { $ref: "#/components/schemas/Decimal" }
        output_unit_price: { $ref: "#/components/schemas/Decimal" }
        price_effective_at: { $ref: "#/components/schemas/DateTime" }
        source_payload: { $ref: "#/components/schemas/JsonValue" }
...